Extract rainbow brush
